requirements.txt
streamlit
pandas
gspread
google-auth
google-auth-oauthlib

